fix(agents): break terminal outcome import cycle
